==> Data/1740_Catastici/index/pipeline/1-SetUp_Dataset.ipynb <==
import pandas as pd
import os

-------------------------------------------------------------------------------------------------------------------

# Dataset setup and unique ID assignment

-------------------------------------------------------------------------------------------------------------------

#### Load dataset

dataset_DF = pd.read_json('../data_catastici/data_pre-processing/catastici_20240123.json')
dataset_DF

#### Insert and assign UIDX column

dataset_DF.insert(0, "uidx", -1)
dataset_DF['uidx'] = dataset_DF.reset_index().index

#### ID uniqueness check

print("Column UID is unique:", dataset_DF['uid'].value_counts().values[0] == 1)
print("Column UIDX is unique:", dataset_DF['uidx'].value_counts().values[0] == 1)

#### Insert columns for new standardised data structure

insert_idx = 3

dataset_DF.insert(insert_idx, "owner_mestiere_std", '')
dataset_DF.insert(insert_idx, "owner_mestiere", '')
dataset_DF.insert(insert_idx, "owner_title_std", '')
dataset_DF.insert(insert_idx, "owner_title", '')
dataset_DF.insert(insert_idx, "owner_family_group", '')
dataset_DF.insert(insert_idx, "owner_family_name", '')
dataset_DF.insert(insert_idx, "owner_first_name", '')
dataset_DF.insert(insert_idx, "owner_entity_group", '')
dataset_DF.insert(insert_idx, "owner_entity", '')
dataset_DF.insert(insert_idx, "owner_count_remark", '')
dataset_DF.insert(insert_idx, "owner_count", '')
dataset_DF.insert(insert_idx, "owner_code", 'TODO')
dataset_DF.head()

### ⚠️⚠️⚠️ Write the updated dataset to file

directory = '../data_catastici/data_post-processing/pipeline_steps'

if not os.path.exists(directory):
    os.makedirs(directory)

dataset_DF.to_json('../data_catastici/data_post-processing/pipeline_steps/catastici_1741_step1.json', orient='records', index=['uidx'])

-----------------------------------------------------------------------------

import time
print("Current time:", time.strftime("%H:%M:%S", time.localtime()))
